requirements.txt
coolsms-python-sdk==2.0.3
Django==3.2.5
django-bootstrap5==2.1.1
requests==2.26.0
pybase62==0.4.3
sentry-sdk==1.3.0
